python-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
